isolate bytes in SchemaAwareJavaBeans and add tests for read and write, refactor test class names to match schematransform name pattern
