+++ sdks/java/io/file-schema-transform/src/test/java/org/apache/beam/sdk/io/fileschematransform/AvroReadSchemaTransformFormatProviderTest.java
@@ -19,22 +19,28 @@ package org.apache.beam.sdk.io.fileschematransform;
 
 import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.ALL_PRIMITIVE_DATA_TYPES_SCHEMA;
 import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.ARRAY_PRIMITIVE_DATA_TYPES_SCHEMA;
+import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.BYTE_SEQUENCE_TYPE_SCHEMA;
+import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.BYTE_TYPE_SCHEMA;
 import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.NULLABLE_ALL_PRIMITIVE_DATA_TYPES_SCHEMA;
 import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.SINGLY_NESTED_DATA_TYPES_SCHEMA;
 import static org.apache.beam.sdk.io.common.SchemaAwareJavaBeans.TIME_CONTAINING_SCHEMA;
 import static org.apache.beam.sdk.io.fileschematransform.FileWriteSchemaTransformFormatProviderTestData.DATA;
-import static org.apache.beam.sdk.transforms.Contextful.fn;
+import static org.apache.beam.sdk.io.fs.ResolveOptions.StandardResolveOptions.RESOLVE_FILE;
 
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.avro.generic.GenericRecord;
-import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.extensions.avro.coders.AvroGenericCoder;
+import org.apache.beam.sdk.extensions.avro.io.AvroIO;
+import org.apache.beam.sdk.extensions.avro.io.DynamicAvroDestinations;
 import org.apache.beam.sdk.extensions.avro.schemas.utils.AvroUtils;
-import org.apache.beam.sdk.io.FileIO;
+import org.apache.beam.sdk.io.DefaultFilenamePolicy;
+import org.apache.beam.sdk.io.FileBasedSink;
+import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.io.GenerateSequence;
-import org.apache.beam.sdk.io.parquet.ParquetIO;
+import org.apache.beam.sdk.io.fs.ResourceId;
+import org.apache.beam.sdk.options.ValueProvider.StaticValueProvider;
 import org.apache.beam.sdk.schemas.Schema;
 import org.apache.beam.sdk.schemas.Schema.Field;
 import org.apache.beam.sdk.schemas.Schema.FieldType;
@@ -58,7 +64,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-public class ParquetFileReadSchemaTransformFormatProviderTest {
+public class AvroReadSchemaTransformFormatProviderTest {
 
   @Rule public TestPipeline writePipeline = TestPipeline.create();
   @Rule public TestPipeline readPipeline = TestPipeline.create();
@@ -70,13 +76,13 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
 
   private String folder(String testName) {
     try {
-      return tmpFolder.newFolder("parquet", testName).getAbsolutePath();
+      return tmpFolder.newFolder("avro", testName).getAbsolutePath();
     } catch (IOException e) {
       throw new IllegalStateException(e);
     }
   }
 
-  public void runWriteAndReadTest(Schema schema, List<Row> rows, String folderPath) {
+  public void runWriteAndReadTest(Schema schema, List<Row> rows, String filePath) {
     org.apache.avro.Schema avroSchema = AvroUtils.toAvroSchema(schema);
     String stringSchema = avroSchema.toString();
 
@@ -86,19 +92,14 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
             MapElements.into(TypeDescriptor.of(GenericRecord.class))
                 .via(AvroUtils.getRowToGenericRecordFunction(avroSchema)))
         .setCoder(AvroGenericCoder.of(avroSchema))
-        .apply(
-            FileIO.<GenericRecord>write()
-                .via(ParquetIO.sink(avroSchema))
-                .to(folderPath)
-                .withSuffix(".parquet"));
+        .apply(AvroIO.writeGenericRecords(avroSchema).to(filePath));
     writePipeline.run().waitUntilFinish();
 
     FileReadSchemaTransformConfiguration config =
         FileReadSchemaTransformConfiguration.builder()
-            .setFormat("parquet")
+            .setFormat("avro")
             .setSchema(stringSchema)
-            // FileIO write with sink writes to a directory
-            .setFilepattern(folderPath + "/*")
+            .setFilepattern(filePath + "*")
             .build();
 
     SchemaTransform readTransform = new FileReadSchemaTransformProvider().from(config);
@@ -132,6 +133,24 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
     Schema schema = TIME_CONTAINING_SCHEMA;
     List<Row> rows = DATA.timeContainingRows;
     String filePath = filePath("testTimeContaining");
+
+    runWriteAndReadTest(schema, rows, filePath);
+  }
+
+  @Test
+  public void testByteType() {
+    Schema schema = BYTE_TYPE_SCHEMA;
+    List<Row> rows = DATA.byteTypeRows;
+    String filePath = filePath("testByteType");
+
+    runWriteAndReadTest(schema, rows, filePath);
+  }
+
+  @Test
+  public void testByteSequenceType() {
+    Schema schema = BYTE_SEQUENCE_TYPE_SCHEMA;
+    List<Row> rows = DATA.byteSequenceTypeRows;
+    String filePath = filePath("testByteSequenceType");
 
     runWriteAndReadTest(schema, rows, filePath);
   }
@@ -175,12 +194,13 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
     List<Row> rows = DATA.allPrimitiveDataTypesRows;
 
     String folder = folder("testStreamingRead");
+    ResourceId dir = FileSystems.matchNewResource(folder, true);
 
     org.apache.avro.Schema avroSchema = AvroUtils.toAvroSchema(schema);
     String stringSchema = avroSchema.toString();
     FileReadSchemaTransformConfiguration config =
         FileReadSchemaTransformConfiguration.builder()
-            .setFormat("parquet")
+            .setFormat("avro")
             .setFilepattern(folder + "/test_*")
             .setSchema(stringSchema)
             .setPollIntervalMillis(100)
@@ -203,13 +223,11 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
         .apply(MapElements.via(new CreateAvroPrimitiveGenericRecord(schema)))
         .setCoder(AvroGenericCoder.of(avroSchema))
         .apply(
-            FileIO.<String, GenericRecord>writeDynamic()
-                .by(fn((GenericRecord element) -> element.get("anInteger").toString()))
-                .via(ParquetIO.sink(avroSchema))
-                .to(folder)
-                .withNaming(integer -> FileIO.Write.defaultNaming("test_" + integer, ".parquet"))
-                .withDestinationCoder(StringUtf8Coder.of())
-                .withNumShards(1));
+            AvroIO.writeGenericRecords(avroSchema)
+                .to(new TestDynamicDestinations(dir))
+                .withTempDirectory(dir)
+                .withNumShards(1)
+                .withWindowedWrites());
 
     // Check output matches the expected rows
     PAssert.that(output.get(FileReadSchemaTransformProvider.OUTPUT_TAG)).containsInAnyOrder(rows);
@@ -224,6 +242,7 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
     String folder = folder("testReadWithPCollectionOfFilepatterns");
 
     org.apache.avro.Schema avroSchema = AvroUtils.toAvroSchema(schema);
+    ResourceId dir = FileSystems.matchNewResource(folder, true);
     // Write rows to dynamic destinations (test_1.., test_2.., test_3..)
     writePipeline
         .apply(Create.of(rows).withRowSchema(schema))
@@ -232,19 +251,16 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
                 .via(AvroUtils.getRowToGenericRecordFunction(avroSchema)))
         .setCoder(AvroGenericCoder.of(avroSchema))
         .apply(
-            FileIO.<String, GenericRecord>writeDynamic()
-                .by(fn((GenericRecord element) -> element.get("anInteger").toString()))
-                .via(ParquetIO.sink(avroSchema))
-                .to(folder)
-                .withNaming(integer -> FileIO.Write.defaultNaming("test_" + integer, ".parquet"))
-                .withDestinationCoder(StringUtf8Coder.of()));
+            AvroIO.writeGenericRecords(avroSchema)
+                .to(new TestDynamicDestinations(dir))
+                .withTempDirectory(dir));
     writePipeline.run().waitUntilFinish();
 
     // We will get filepatterns from the input PCollection, so don't set filepattern field here
     String stringSchema = avroSchema.toString();
     FileReadSchemaTransformConfiguration config =
         FileReadSchemaTransformConfiguration.builder()
-            .setFormat("parquet")
+            .setFormat("avro")
             .setSchema(stringSchema)
             .build();
     SchemaTransform readTransform = new FileReadSchemaTransformProvider().from(config);
@@ -274,5 +290,45 @@ public class ParquetFileReadSchemaTransformFormatProviderTest {
     // Check output matches with expected rows
     PAssert.that(output.get(FileReadSchemaTransformProvider.OUTPUT_TAG)).containsInAnyOrder(rows);
     readPipeline.run();
+  }
+
+  private static class TestDynamicDestinations
+      extends DynamicAvroDestinations<GenericRecord, String, GenericRecord> {
+    final ResourceId baseDir;
+
+    TestDynamicDestinations(ResourceId baseDir) {
+      this.baseDir = baseDir;
+    }
+
+    @Override
+    public org.apache.avro.Schema getSchema(String destination) {
+      return AvroUtils.toAvroSchema(ALL_PRIMITIVE_DATA_TYPES_SCHEMA);
+    }
+
+    @Override
+    public GenericRecord formatRecord(GenericRecord record) {
+      return record;
+    }
+
+    @Override
+    public String getDestination(GenericRecord element) {
+      // Destination will be either test_1, test_2, or test_3 depending on the value of
+      // anInteger field.
+      return element.get("anInteger").toString();
+    }
+
+    @Override
+    public String getDefaultDestination() {
+      return "";
+    }
+
+    @Override
+    public FileBasedSink.FilenamePolicy getFilenamePolicy(String destination) {
+      return DefaultFilenamePolicy.fromStandardParameters(
+          StaticValueProvider.of(baseDir.resolve("test_" + destination, RESOLVE_FILE)),
+          "-SSSSS-of-NNNNN",
+          ".avro",
+          false);
+    }
   }
 }
