+++ sdks/java/io/file-schema-transform/src/test/java/org/apache/beam/sdk/io/fileschematransform/FileReadSchemaTransformFormatProviderTest.java
@@ -1,0 +1,5 @@
+package org.apache.beam.sdk.io.fileschematransform;
+
+public class FileReadSchemaTransformFormatProviderTest {
+
+}
